vortex.bootstrap: initial (prototype) stage 1 bootstrap implementation
This is a bit rough around the edges and needs far more error checking,
as well as testing on the target systems. Works on Debian 8.x with
Python 2 and Python 3.

Reads /etc/vortex.ini and obtains the main Vortex package from the
configured URL. This is expected to be a Python Egg, which is added to
Python's module path, and the entry point is executed.