vortex.config: add absorb() and co, don't try to validate
The config module has no business knowing what other modules need to get
from the config, so let's delegate that to the modules themselves. To
help with that, provide an absorb() method which will check for required
configuration options, set default values for optional ones, and then
setattr() the values into a target object.