vortex.deployment: add API docs, pass deployer to handlers
Add all the requisite API documentation to this module. Also pass the
deployer to handler objects so they can obtain more information from the
payload (e.g. the payload directory). This will be useful for the exec
handler so it can run scripts with the payload directory as the working
directory, and so it can pass through the "environment" configuration
parameter.